+++ src/Analysis/Relations/TraitUsageResolver.php
@@ -38,9 +38,12 @@ class TraitUsageResolver extends AbstractResolver
                     ($traitAlias->getTraitFqcn() === null  || $traitAlias->getTraitFqcn() === $trait['fqcn'])
                 ) {
                     $method['name']        = $traitAlias->getAlias() ?: $method['name'];
-                    $method['isPublic']    = ($traitAlias->getAccessModifier()->getName() === AccessModifierNameValue::PUBLIC_);
-                    $method['isProtected'] = ($traitAlias->getAccessModifier()->getName() === AccessModifierNameValue::PROTECTED_);
-                    $method['isPrivate']   = ($traitAlias->getAccessModifier()->getName() === AccessModifierNameValue::PRIVATE_);
+
+                    if ($traitAlias->getAccessModifier()) {
+                        $method['isPublic']    = ($traitAlias->getAccessModifier()->getName() === AccessModifierNameValue::PUBLIC_);
+                        $method['isProtected'] = ($traitAlias->getAccessModifier()->getName() === AccessModifierNameValue::PROTECTED_);
+                        $method['isPrivate']   = ($traitAlias->getAccessModifier()->getName() === AccessModifierNameValue::PRIVATE_);
+                    }
                 }
             }